requirements.txt
python-telegram-bot[ext]>=20.0
requests
httpx